requirements.txt
SQLAlchemy
openai
pdfminer.six
goose3
sumy
selenium
psycopg2
redis==5.0.8
anyio==4.2.0
APScheduler==3.10.4
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
lxml==4.9.3
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tornado==6.3.3
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
dateparser==1.2.0
